requirements.txt
fuzzywuzzy==0.18.0
minio==7.2.5
numpy==1.21.6
numpy==1.23.5
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
paddleocr==2.6.0.1
paddleocr.egg==info
pandas==2.2.1
Pillow==10.3.0
PyQt5==5.15.10
PyQt5_sip==12.13.0
pytesseract==0.3.10
Requests==2.31.0
wget==3.2
